Update Zotero8 fork docs and release flow
