+++ package.json
@@ -5,12 +5,15 @@
   "author": "Jose",
   "license": "GPL-3.0",
   "scripts": {
+    "bundle": "node build/bundle.mjs",
     "build": "node build/build.mjs",
+    "release": "npm run build && npm run bundle && npm run package && npm run release:upload",
+    "release:upload": "node build/release-upload.mjs",
     "package": "node build/package.mjs",
-    "release": "npm run build && npm run package",
     "clean": "rm -rf dist *.xpi"
   },
   "devDependencies": {
-    "archiver": "^7.0.0"
+    "archiver": "^7.0.0",
+    "esbuild": "^0.28.0"
   }
 }
